+++ src/lib/barrels.ts
@@ -14,6 +14,7 @@
  * limitations under the License.
  */
 import { PipesModule } from './pipes/pipes.module';
+import { StAlertsModule } from './st-alerts/st-alerts.module';
 import { StBreadcrumbsModule } from './st-breadcrumbs/st-breadcrumbs.module';
 import { StButtonModule } from './st-button/st-button.module';
 import { StCheckboxModule } from './st-checkbox/st-checkbox.module';
@@ -50,6 +51,7 @@ import { StFormDirectiveModule } from './directives/form/form-directives.module'
 
 export const DECLARATIONS: any[] = [
    PipesModule,
+   StAlertsModule,
    StBreadcrumbsModule,
    StButtonModule,
    StCheckboxModule,