+++ src/lib/egeo.module.ts
@@ -16,6 +16,7 @@
 import { CommonModule } from '@angular/common';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 
+import { StAlertsService } from './st-alerts/st-alerts.service';
 import { StModalService } from './st-modal/st-modal.service';
 import { StPaginationService } from './st-pagination/st-pagination.service';
 import { EgeoResolveService } from './utils/egeo-resolver/egeo-resolve.service';
@@ -37,7 +38,7 @@ export class EgeoModule {
    static forRoot(): ModuleWithProviders {
       return {
          ngModule: EgeoModule,
-         providers: [StModalService, StPaginationService, EgeoResolveService, SelectOneDispatcher]
+         providers: [StModalService, StPaginationService, EgeoResolveService, SelectOneDispatcher, StAlertsService]
       };
    }
 }