+++ src/lib/st-alerts/st-alerts.component.ts
@@ -13,21 +13,24 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-export interface MenuOption {
-   route: string;
-   name: string;
-}
+import { Component, Input, ChangeDetectionStrategy, ChangeDetectorRef } from '@angular/core';
 
-export const OPTIONS: MenuOption[] = [
-   { route: 'main', name: 'main' },
-   { route: 'alerts-demo', name: 'Alerts' },
-   { route: 'dropdown-menu-demo', name: 'Dropdown Menu' },
-   { route: 'dropdown-demo', name: 'Dropdown' },
-   { route: 'combobox-demo', name: 'Combobox' },
-   { route: 'help-demo', name: 'help' },
-   { route: 'info-box-demo', name: 'info-box' },
-   { route: 'item-list-demo', name: 'item-list' },
-   { route: 'tip-demo', name: 'tip' },
-   { route: 'search-demo', name: 'Search' },
-   { route: 'tree-demo', name: 'Tree' }
-];
+import { StAlert } from './st-alerts.model';
+import { StAlertsService } from './st-alerts.service';
+
+@Component({
+   selector: 'st-alerts',
+   templateUrl: './st-alerts.component.html',
+   styleUrls: ['./st-alerts.component.scss'],
+   changeDetection: ChangeDetectionStrategy.OnPush
+})
+export class StAlertsComponent {
+
+   @Input() showInConsole: boolean = false;
+   @Input() qaTag: string = 'st-alert-manager';
+
+   constructor(
+      private _cd: ChangeDetectorRef,
+      public alertService: StAlertsService
+   ) { }
+}